+++ src/examples/zkapps/escrow/escrow.ts
@@ -1,13 +1,7 @@
-import {
-  SmartContract,
-  method,
-  UInt64,
-  AccountUpdate,
-  PublicKey,
-} from 'o1js';
+import { SmartContract, method, UInt64, AccountUpdate, PublicKey } from 'o1js';
 
 export class Escrow extends SmartContract {
-  @method deposit(user: PublicKey) {
+  @method async deposit(user: PublicKey) {
     // add your deposit logic circuit here
     // that will adjust the amount
 
@@ -15,7 +9,7 @@ export class Escrow extends SmartContract {
     payerUpdate.send({ to: this.address, amount: UInt64.from(1000000) });
   }
 
-  @method withdraw(user: PublicKey) {
+  @method async withdraw(user: PublicKey) {
     // add your withdrawal logic circuit here
     // that will adjust the amount