+++ src/examples/zkapps/reducer/map.ts
@@ -12,6 +12,7 @@ import {
   Bool,
   Poseidon,
   Provable,
+  assert,
 } from 'o1js';
 
 /*
@@ -50,11 +51,12 @@ class StorageContract extends SmartContract {
     actionType: KeyValuePair,
   });
 
-  @method set(key: PublicKey, value: Field) {
+  @method async set(key: PublicKey, value: Field) {
     this.reducer.dispatch({ key: Poseidon.hash(key.toFields()), value });
   }
 
-  @method get(key: PublicKey): Option {
+  @method.returns(Option)
+  async get(key: PublicKey) {
     let pendingActions = this.reducer.getActions({
       fromActionState: Reducer.initialActionState,
     });
@@ -105,7 +107,7 @@ await StorageContract.compile();
 
 let tx = await Mina.transaction(feePayer, async () => {
   AccountUpdate.fundNewAccount(feePayer);
-  zkapp.deploy();
+  await zkapp.deploy();
 });
 await tx.sign([feePayerKey, zkappKey]).send();
 
@@ -131,7 +133,7 @@ let value = map[0].value;
 console.log(`setting key ${key.toBase58()} with value ${value}`);
 
 tx = await Mina.transaction(feePayer, async () => {
-  zkapp.set(key, value);
+  await zkapp.set(key, value);
 });
 await tx.prove();
 await tx.sign([feePayerKey]).send();
@@ -141,7 +143,7 @@ value = map[1].value;
 console.log(`setting key ${key.toBase58()} with value ${value}`);
 
 tx = await Mina.transaction(feePayer, async () => {
-  zkapp.set(key, value);
+  await zkapp.set(key, value);
 });
 await tx.prove();
 await tx.sign([feePayerKey]).send();
@@ -151,7 +153,7 @@ value = map[2].value;
 console.log(`setting key ${key.toBase58()} with value ${value}`);
 
 tx = await Mina.transaction(feePayer, async () => {
-  zkapp.set(key, value);
+  await zkapp.set(key, value);
 });
 await tx.prove();
 await tx.sign([feePayerKey]).send();
@@ -160,13 +162,14 @@ key = map[0].key;
 value = map[0].value;
 console.log(`getting key ${key.toBase58()} with value ${value}`);
 
-let result: any;
+let result: Option | undefined;
 tx = await Mina.transaction(feePayer, async () => {
-  result = zkapp.get(key);
+  result = await zkapp.get(key);
 });
 await tx.prove();
 await tx.sign([feePayerKey]).send();
 
+assert(result !== undefined);
 console.log('found correct match?', result.isSome.toBoolean());
 console.log('matches expected value?', result.value.equals(value).toBoolean());
 
@@ -175,7 +178,7 @@ value = map[1].value;
 console.log(`getting key ${key.toBase58()} with value ${value}`);
 
 tx = await Mina.transaction(feePayer, async () => {
-  result = zkapp.get(key);
+  result = await zkapp.get(key);
 });
 await tx.prove();
 await tx.sign([feePayerKey]).send();
@@ -185,7 +188,7 @@ console.log('matches expected value?', result.value.equals(value).toBoolean());
 
 console.log(`getting key invalid key`);
 tx = await Mina.transaction(feePayer, async () => {
-  result = zkapp.get(PrivateKey.random().toPublicKey());
+  result = await zkapp.get(PrivateKey.random().toPublicKey());
 });
 await tx.prove();
 await tx.sign([feePayerKey]).send();