+++ src/examples/zkapps/reducer/reducer-composite.ts
@@ -31,14 +31,14 @@ class CounterZkapp extends SmartContract {
   // helper field to store the point in the action history that our on-chain state is at
   @state(Field) actionState = State<Field>();
 
-  @method incrementCounter() {
+  @method async incrementCounter() {
     this.reducer.dispatch(INCREMENT);
   }
-  @method dispatchData(data: Field) {
+  @method async dispatchData(data: Field) {
     this.reducer.dispatch({ isIncrement: Bool(false), otherData: data });
   }
 
-  @method rollupIncrements() {
+  @method async rollupIncrements() {
     // get previous counter & actions hash, assert that they're the same as on-chain values
     let counter = this.counter.get();
     this.counter.requireEquals(counter);
@@ -94,7 +94,7 @@ if (doProofs) {
 console.log('deploy');
 let tx = await Mina.transaction(feePayer, async () => {
   AccountUpdate.fundNewAccount(feePayer);
-  zkapp.deploy();
+  await zkapp.deploy();
   zkapp.counter.set(initialCounter);
   zkapp.actionState.set(Reducer.initialActionState);
 });
@@ -105,21 +105,21 @@ console.log('applying actions..');
 console.log('action 1');
 
 tx = await Mina.transaction(feePayer, async () => {
-  zkapp.incrementCounter();
+  await zkapp.incrementCounter();
 });
 await tx.prove();
 await tx.sign([feePayerKey]).send();
 
 console.log('action 2');
 tx = await Mina.transaction(feePayer, async () => {
-  zkapp.incrementCounter();
+  await zkapp.incrementCounter();
 });
 await tx.prove();
 await tx.sign([feePayerKey]).send();
 
 console.log('action 3');
 tx = await Mina.transaction(feePayer, async () => {
-  zkapp.incrementCounter();
+  await zkapp.incrementCounter();
 });
 await tx.prove();
 await tx.sign([feePayerKey]).send();
@@ -129,7 +129,7 @@ console.log('rolling up pending actions..');
 console.log('state before: ' + zkapp.counter.get());
 
 tx = await Mina.transaction(feePayer, async () => {
-  zkapp.rollupIncrements();
+  await zkapp.rollupIncrements();
 });
 await tx.prove();
 await tx.sign([feePayerKey]).send();
@@ -141,14 +141,14 @@ console.log('applying more actions');
 
 console.log('action 4 (no increment)');
 tx = await Mina.transaction(feePayer, async () => {
-  zkapp.dispatchData(Field.random());
+  await zkapp.dispatchData(Field.random());
 });
 await tx.prove();
 await tx.sign([feePayerKey]).send();
 
 console.log('action 5');
 tx = await Mina.transaction(feePayer, async () => {
-  zkapp.incrementCounter();
+  await zkapp.incrementCounter();
 });
 await tx.prove();
 await tx.sign([feePayerKey]).send();
@@ -158,7 +158,7 @@ console.log('rolling up pending actions..');
 console.log('state before: ' + zkapp.counter.get());
 
 tx = await Mina.transaction(feePayer, async () => {
-  zkapp.rollupIncrements();
+  await zkapp.rollupIncrements();
 });
 await tx.prove();
 await tx.sign([feePayerKey]).send();