+++ src/examples/zkapps/voting/voting.ts
@@ -97,8 +97,8 @@ export class Voting_ extends SmartContract {
     }),
   };
 
-  deploy(args: DeployArgs) {
-    super.deploy(args);
+  async deploy(args: DeployArgs) {
+    await super.deploy(args);
     this.account.permissions.set({
       ...Permissions.default(),
       editState: Permissions.proofOrSignature(),
@@ -118,7 +118,7 @@ export class Voting_ extends SmartContract {
    * @param member
    */
   @method
-  voterRegistration(member: Member) {
+  async voterRegistration(member: Member) {
     let currentSlot = this.network.globalSlotSinceGenesis.get();
     this.network.globalSlotSinceGenesis.requireBetween(
       currentSlot,
@@ -153,7 +153,7 @@ export class Voting_ extends SmartContract {
     );
 
     let VoterContract: Membership_ = new Membership_(voterAddress);
-    let exists = VoterContract.addEntry(member);
+    let exists = await VoterContract.addEntry(member);
 
     // the check happens here because we want to see if the other contract returns a value
     // if exists is true, that means the member already exists within the accumulated state
@@ -167,7 +167,7 @@ export class Voting_ extends SmartContract {
    * @param member
    */
   @method
-  candidateRegistration(member: Member) {
+  async candidateRegistration(member: Member) {
     let currentSlot = this.network.globalSlotSinceGenesis.get();
     this.network.globalSlotSinceGenesis.requireBetween(
       currentSlot,
@@ -202,7 +202,7 @@ export class Voting_ extends SmartContract {
     );
 
     let CandidateContract: Membership_ = new Membership_(candidateAddress);
-    let exists = CandidateContract.addEntry(member);
+    let exists = await CandidateContract.addEntry(member);
 
     // the check happens here because we want to see if the other contract returns a value
     // if exists is true, that means the member already exists within the accumulated state
@@ -215,13 +215,13 @@ export class Voting_ extends SmartContract {
    * Calls the `publish()` method of the Candidate-Membership and Voter-Membership contract.
    */
   @method
-  approveRegistrations() {
+  async approveRegistrations() {
     // Invokes the publish method of both Voter and Candidate Membership contracts.
     let VoterContract: Membership_ = new Membership_(voterAddress);
-    VoterContract.publish();
+    await VoterContract.publish();
 
     let CandidateContract: Membership_ = new Membership_(candidateAddress);
-    CandidateContract.publish();
+    await CandidateContract.publish();
   }
 
   /**
@@ -231,7 +231,7 @@ export class Voting_ extends SmartContract {
    * @param voter
    */
   @method
-  vote(candidate: Member, voter: Member) {
+  async vote(candidate: Member, voter: Member) {
     let currentSlot = this.network.globalSlotSinceGenesis.get();
     this.network.globalSlotSinceGenesis.requireBetween(
       currentSlot,
@@ -250,10 +250,10 @@ export class Voting_ extends SmartContract {
     // verifying that both the voter and the candidate are actually part of our member set
     // ideally we would also verify a signature here, but ignoring that for now
     let VoterContract: Membership_ = new Membership_(voterAddress);
-    VoterContract.isMember(voter).assertTrue('Member is not a voter!');
+    (await VoterContract.isMember(voter)).assertTrue('Member is not a voter!');
 
     let CandidateContract: Membership_ = new Membership_(candidateAddress);
-    CandidateContract.isMember(candidate).assertTrue(
+    (await CandidateContract.isMember(candidate)).assertTrue(
       'Member is not a candidate!'
     );
 
@@ -268,7 +268,7 @@ export class Voting_ extends SmartContract {
    * and applies state changes to the votes merkle tree.
    */
   @method
-  countVotes() {
+  async countVotes() {
     let accumulatedVotes = this.accumulatedVotes.get();
     this.accumulatedVotes.requireEquals(accumulatedVotes);