+++ src/lib/util/assert.ts
@@ -6,7 +6,7 @@ function assert(stmt: boolean, message?: string): asserts stmt {
   }
 }
 
-function assertPromise<T>(value: Promise<T>): Promise<T> {
-  assert(value instanceof Promise, 'Expected a promise');
+function assertPromise<T>(value: Promise<T>, message?: string): Promise<T> {
+  assert(value instanceof Promise, message ?? 'Expected a promise');
   return value;
 }